program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line. Updated

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i;
-	long n1 = 1, n2 = 2, n3;
+	float n1 = 1, n2 = 2, n3;
 
 	printf("1, 2, ");
 	for (i = 2; i < 98; i++)
 	{
 		n3 = n1 + n2;
-		printf("%ld", n3);
+		printf("%.0f", n3);
 		n1 = n2;
 		n2 = n3;
 		if (i == (98 - 1))
